requirements.txt
python-telegram-bot
requests
beautifulsoup4
pandas
pyarrow
pdfplumber
openai